Move schemas and types
