+++ server/types/hubManager.ts
@@ -1,0 +1,6 @@
+type HubManager = {
+  name: string
+  hasSignature: boolean
+}
+
+export default HubManager
